psql: Fix pager setup for unaligned and expanded output
Move the line counting logic to the pager setup so that it applies to
unaligned and expanded output modes as well.  While at it, get rid of
the top if-else branch in IsPagerNeeded in favor of an early return when
not writing to stdout.  This avoids having extra indentation for the
most part of that function.