psql: Fix counting of cell lines in expanded mode
In expanded mode we repeat the table header in every record by printing
header and cell side by side.  But the pager setup only counts the lines
in the cells.  The total line count is off if headers have more line
breaks than some of the respective cells.  Fix this by distinguishing
between expanded and normal mode in the loop that counts the cell lines.
While at it, change that loop to count all lines of each cell instead of
just the extra lines to make the logic easier to follow.  The first line
of each cell was accounted for in the initial value of the total line
count which now only accounts for the separator lines, depending on the
output mode.